refactor: Extract connection helper to transaction query helper file
